Passage du double en float

+++ Bernstein.cc
@@ -21,9 +21,9 @@ int coef_binomial(int n, int k)
     return factorielle(n) / (factorielle(k) * factorielle(n - k));
 }
 
-double bernstein(int degre, int indice, double t)
+float bernstein(int degre, int indice, float t)
 {
-    double resultat = coef_binomial(degre, indice) * pow(t, indice) * pow(1 - t, degre - indice);
+    float resultat = coef_binomial(degre, indice) * pow(t, indice) * pow(1 - t, degre - indice);
 
     return resultat;
 }
