+++ Dessin.cc
@@ -10,7 +10,7 @@ using namespace std;
  ****************************************************************/
 
 // Trace un point à partir de coordonnées
-void trace_point(double x, double y, double red, double green, double blue, double size)
+void trace_point(float x, float y, float red, float green, float blue, float size)
 {
     glColor3f(red, green, blue); // initialisation de la couleur
     glPointSize(size);           // initialisation de la taille
@@ -20,13 +20,13 @@ void trace_point(double x, double y, double red, double green, double blue, doub
 }
 
 // Trace un point à partir d'un objet Point
-void trace_point(Point p, double red, double green, double blue, double size)
+void trace_point(Point p, float red, float green, float blue, float size)
 {
     trace_point(p.getX(), p.getY(), red, green, blue, size);
 }
 
 // Trace un segement à partir des coordonnées des points
-void trace_segment(double x0, double y0, double x1, double y1, double red, double green, double blue, double size)
+void trace_segment(float x0, float y0, float x1, float y1, float red, float green, float blue, float size)
 {
     glColor3f(red, green, blue); // initialisation de la couleur
     glLineWidth(size);           // initialisation de la taille
@@ -37,13 +37,13 @@ void trace_segment(double x0, double y0, double x1, double y1, double red, doubl
 }
 
 // Trace un segment à partir de deux objets points
-void trace_segment(Point p0, Point p1, double red, double green, double blue, double size)
+void trace_segment(Point p0, Point p1, float red, float green, float blue, float size)
 {
     trace_segment(p0.getX(), p0.getY(), p1.getX(), p1.getY(), red, green, blue, size);
 }
 
 // Trace un segment à partir d'un objet segment
-void trace_segment(Segment s, double red, double green, double blue, double size)
+void trace_segment(Segment s, float red, float green, float blue, float size)
 {
     trace_segment(s.getA(), s.getB(), red, green, blue, size);
 }
