+++ Polynome.h
@@ -13,20 +13,20 @@ class Polynome
     Polynome(int degre);
     Polynome(const Polynome &original);
     Polynome(const Polynome &original, int degre_limite);
-    Polynome(std::vector<double> coefficients);
-    Polynome(double coefficients[], size_t taille);
+    Polynome(std::vector<float> coefficients);
+    Polynome(float coefficients[], size_t taille);
 
     // Méthodes
     void normalize();
     int get_degre() const;
-    int solve(double &x1, double &x2) const;
-    double get_last_coef() const;
-    double get_coef_puissance(int d) const;
-    void add_coef(double c);
-    double value_for(double x) const;
-    std::vector<double> get_coefs() const;
+    int solve(float &x1, float &x2) const;
+    float get_last_coef() const;
+    float get_coef_puissance(int d) const;
+    void add_coef(float c);
+    float value_for(float x) const;
+    std::vector<float> get_coefs() const;
     static Polynome div_euclide(Polynome const &p1, Polynome const &p2);
-    std::vector<double> passage_base_bernstein(int degre) const;
+    std::vector<float> passage_base_bernstein(int degre) const;
 
     //Surchage d'opérateur
     Polynome &operator+=(const Polynome &p);
@@ -38,10 +38,10 @@ class Polynome
     void affiche_polynome();
     void affiche_polynome(std::string message);
 
-    void decalage(int puissance, double coefficient);
+    void decalage(int puissance, float coefficient);
   private:
 
-    std::vector<double> m_coefficients;
+    std::vector<float> m_coefficients;
 
 };
